Rename API gateway package
