+++ api_gateway/Dockerfile
@@ -19,6 +19,6 @@ COPY --from=build ${DEPENDENCY}/BOOT-INF/lib /app/lib
 COPY --from=build ${DEPENDENCY}/META-INF /app/META-INF
 COPY --from=build ${DEPENDENCY}/BOOT-INF/classes /app
 
-ENTRYPOINT ["java", "-cp", "app:app/lib/*", "pl.edu.pg.rsww.api_gateway.ApiGatewayApplicationKt"]
+ENTRYPOINT ["java", "-cp", "app:app/lib/*", "pl.edu.pg.rsww.apigateway.ApiGatewayApplicationKt"]
 
 HEALTHCHECK CMD wget -O - http://127.0.0.1:8080/api || exit 1
