+++ api_gateway/src/main/kotlin/pl/edu/pg/rsww/apigateway/NotFoundException.kt
@@ -1,4 +1,4 @@
-package pl.edu.pg.rsww.api_gateway
+package pl.edu.pg.rsww.apigateway
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
