+++ api_gateway/src/main/kotlin/pl/edu/pg/rsww/apigateway/RequestMessage.kt
@@ -1,4 +1,4 @@
-package pl.edu.pg.rsww.api_gateway
+package pl.edu.pg.rsww.apigateway
 
 import kotlinx.serialization.Serializable
