+++ api_gateway/src/test/kotlin/pl/edu/pg/rsww/apigateway/ApiGatewayApplicationTests.kt
@@ -1,4 +1,4 @@
-package pl.edu.pg.rsww.api_gateway
+package pl.edu.pg.rsww.apigateway
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
